requirements.txt
altgraph==0.17.2
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
buildozer==1.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
cssselect2==0.5.0
Cython==0.29.26
decouple==0.0.7
distlib==0.3.4
Django==3.0.5
django-admin-tools==0.9.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
docutils==0.18.1
filelock==3.4.2
future==0.18.2
html5lib==1.1
idna==3.3
importlib-metadata==4.10.1
Jinja2==3.0.3
Kivy==2.0.0
Kivy-Garden==0.1.4
lxml==4.8.0
macholib==1.15.2
MarkupSafe==2.0.1
mccabe==0.6.1
oscrypto==1.3.0
pep517==0.6.0
pexpect==4.8.0
Pillow==9.1.0
pipenv==2022.1.8
platformdirs==2.4.1
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
pyobjc==8.2
pyobjc-core==8.2
pyobjc-framework-Accessibility==8.2
pyobjc-framework-Accounts==8.2
pyobjc-framework-AddressBook==8.2
pyobjc-framework-AdServices==8.2
pyobjc-framework-AdSupport==8.2
pyobjc-framework-AppleScriptKit==8.2
pyobjc-framework-AppleScriptObjC==8.2
pyobjc-framework-ApplicationServices==8.2
pyobjc-framework-AppTrackingTransparency==8.2
pyobjc-framework-AudioVideoBridging==8.2
pyobjc-framework-AuthenticationServices==8.2
pyobjc-framework-AutomaticAssessmentConfiguration==8.2
pyobjc-framework-Automator==8.2
pyobjc-framework-AVFoundation==8.2
pyobjc-framework-AVKit==8.2
pyobjc-framework-BusinessChat==8.2
pyobjc-framework-CalendarStore==8.2
pyobjc-framework-CallKit==8.2
pyobjc-framework-CFNetwork==8.2
pyobjc-framework-ClassKit==8.2
pyobjc-framework-CloudKit==8.2
pyobjc-framework-Cocoa==8.2
pyobjc-framework-Collaboration==8.2
pyobjc-framework-ColorSync==8.2
pyobjc-framework-Contacts==8.2
pyobjc-framework-ContactsUI==8.2
pyobjc-framework-CoreAudio==8.2
pyobjc-framework-CoreAudioKit==8.2
pyobjc-framework-CoreBluetooth==8.2
pyobjc-framework-CoreData==8.2
pyobjc-framework-CoreHaptics==8.2
pyobjc-framework-CoreLocation==8.2
pyobjc-framework-CoreMedia==8.2
pyobjc-framework-CoreMediaIO==8.2
pyobjc-framework-CoreMIDI==8.2
pyobjc-framework-CoreML==8.2
pyobjc-framework-CoreMotion==8.2
pyobjc-framework-CoreServices==8.2
pyobjc-framework-CoreSpotlight==8.2
pyobjc-framework-CoreText==8.2
pyobjc-framework-CoreWLAN==8.2
pyobjc-framework-CryptoTokenKit==8.2
pyobjc-framework-DeviceCheck==8.2
pyobjc-framework-DictionaryServices==8.2
pyobjc-framework-DiscRecording==8.2
pyobjc-framework-DiscRecordingUI==8.2
pyobjc-framework-DiskArbitration==8.2
pyobjc-framework-DVDPlayback==8.2
pyobjc-framework-EventKit==8.2
pyobjc-framework-ExceptionHandling==8.2
pyobjc-framework-ExecutionPolicy==8.2
pyobjc-framework-ExternalAccessory==8.2
pyobjc-framework-FileProvider==8.2
pyobjc-framework-FileProviderUI==8.2
pyobjc-framework-FinderSync==8.2
pyobjc-framework-FSEvents==8.2
pyobjc-framework-GameCenter==8.2
pyobjc-framework-GameController==8.2
pyobjc-framework-GameKit==8.2
pyobjc-framework-GameplayKit==8.2
pyobjc-framework-ImageCaptureCore==8.2
pyobjc-framework-IMServicePlugIn==8.2
pyobjc-framework-InputMethodKit==8.2
pyobjc-framework-InstallerPlugins==8.2
pyobjc-framework-InstantMessage==8.2
pyobjc-framework-Intents==8.2
pyobjc-framework-IOSurface==8.2
pyobjc-framework-iTunesLibrary==8.2
pyobjc-framework-KernelManagement==8.2
pyobjc-framework-LatentSemanticMapping==8.2
pyobjc-framework-LaunchServices==8.2
pyobjc-framework-libdispatch==8.2
pyobjc-framework-LinkPresentation==8.2
pyobjc-framework-LocalAuthentication==8.2
pyobjc-framework-MapKit==8.2
pyobjc-framework-MediaAccessibility==8.2
pyobjc-framework-MediaLibrary==8.2
pyobjc-framework-MediaPlayer==8.2
pyobjc-framework-MediaToolbox==8.2
pyobjc-framework-Metal==8.2
pyobjc-framework-MetalKit==8.2
pyobjc-framework-MetalPerformanceShaders==8.2
pyobjc-framework-MetalPerformanceShadersGraph==8.2
pyobjc-framework-MLCompute==8.2
pyobjc-framework-ModelIO==8.2
pyobjc-framework-MultipeerConnectivity==8.2
pyobjc-framework-NaturalLanguage==8.2
pyobjc-framework-NetFS==8.2
pyobjc-framework-Network==8.2
pyobjc-framework-NetworkExtension==8.2
pyobjc-framework-NotificationCenter==8.2
pyobjc-framework-OpenDirectory==8.2
pyobjc-framework-OSAKit==8.2
pyobjc-framework-OSLog==8.2
pyobjc-framework-PassKit==8.2
pyobjc-framework-PencilKit==8.2
pyobjc-framework-Photos==8.2
pyobjc-framework-PhotosUI==8.2
pyobjc-framework-PreferencePanes==8.2
pyobjc-framework-PushKit==8.2
pyobjc-framework-Quartz==8.2
pyobjc-framework-QuickLookThumbnailing==8.2
pyobjc-framework-ReplayKit==8.2
pyobjc-framework-SafariServices==8.2
pyobjc-framework-SceneKit==8.2
pyobjc-framework-ScreenSaver==8.2
pyobjc-framework-ScreenTime==8.2
pyobjc-framework-ScriptingBridge==8.2
pyobjc-framework-SearchKit==8.2
pyobjc-framework-Security==8.2
pyobjc-framework-SecurityFoundation==8.2
pyobjc-framework-SecurityInterface==8.2
pyobjc-framework-ServiceManagement==8.2
pyobjc-framework-Social==8.2
pyobjc-framework-SoundAnalysis==8.2
pyobjc-framework-Speech==8.2
pyobjc-framework-SpriteKit==8.2
pyobjc-framework-StoreKit==8.2
pyobjc-framework-SyncServices==8.2
pyobjc-framework-SystemConfiguration==8.2
pyobjc-framework-SystemExtensions==8.2
pyobjc-framework-UniformTypeIdentifiers==8.2
pyobjc-framework-UserNotifications==8.2
pyobjc-framework-UserNotificationsUI==8.2
pyobjc-framework-VideoSubscriberAccount==8.2
pyobjc-framework-VideoToolbox==8.2
pyobjc-framework-Virtualization==8.2
pyobjc-framework-Vision==8.2
pyobjc-framework-WebKit==8.2
pyotp==2.6.0
PyPDF3==1.0.5
python-bidi==0.4.2
python-decouple==3.6
pytoml==0.1.21
pyttsx3==2.90
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
sh==1.14.2
six==1.16.0
sqlparse==0.3.1
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.62.3
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
webencodings==0.5.1
xhtml2pdf==0.2.7
zipp==3.7.0
